Generated API documentation

The code is generated by the OpenAPI generator with the Julia plugin.

GptSearchPlugin.AppServer.GptPluginServer — Module

Encapsulates generated server code for GptPluginServer

The following server methods must be implemented:

delete_docs

invocation: DELETE /delete

signature: deletedocs(req::HTTP.Request, deleterequest::DeleteRequest;) -> DeleteResponse

query_post

invocation: POST /query

signature: querypost(req::HTTP.Request, queryrequest::QueryRequest;) -> QueryResponse

upsert_post

invocation: POST /upsert

signature: upsertpost(req::HTTP.Request, upsertrequest::UpsertRequest;) -> UpsertResponse

GptSearchPlugin.AppServer.GptPluginServer.DeleteRequest — Type

DeleteRequest

DeleteRequest(;
    ids=nothing,
    filter=nothing,
    delete_all=false,
)

- ids::Vector{String}
- filter::Vector{DocumentMetadataFilter}
- delete_all::Bool

GptSearchPlugin.AppServer.GptPluginServer.DeleteResponse — Type

DeleteResponse

DeleteResponse(;
    success=nothing,
)

- success::Bool

GptSearchPlugin.AppServer.GptPluginServer.Document — Type

Document

Document(;
    id=nothing,
    text=nothing,
    metadata=nothing,
)

- id::String
- text::String
- metadata::DocumentMetadata

GptSearchPlugin.AppServer.GptPluginServer.DocumentChunk — Type

DocumentChunk

DocumentChunk(;
    id=nothing,
    text=nothing,
    metadata=nothing,
    embedding=nothing,
    score=nothing,
)

- id::String
- text::String
- metadata::DocumentChunkMetadata
- embedding::Vector{Float64}
- score::Float64

GptSearchPlugin.AppServer.GptPluginServer.DocumentChunkMetadata — Type

DocumentChunkMetadata

DocumentChunkMetadata(;
    source=nothing,
    source_id=nothing,
    url=nothing,
    created_at=nothing,
    author=nothing,
    document_id=nothing,
)

- source::Source
- source_id::String
- url::String
- created_at::String
- author::String
- document_id::String

GptSearchPlugin.AppServer.GptPluginServer.DocumentChunkWithScore — Type

DocumentChunkWithScore

DocumentChunkWithScore(;
    id=nothing,
    text=nothing,
    metadata=nothing,
    embedding=nothing,
    score=nothing,
)

- id::String
- text::String
- metadata::DocumentChunkMetadata
- embedding::Vector{Float64}
- score::Float64

GptSearchPlugin.AppServer.GptPluginServer.DocumentMetadata — Type

DocumentMetadata

DocumentMetadata(;
    source=nothing,
    source_id=nothing,
    url=nothing,
    created_at=nothing,
    author=nothing,
)

- source::Source
- source_id::String
- url::String
- created_at::String
- author::String

GptSearchPlugin.AppServer.GptPluginServer.DocumentMetadataFilter — Type

DocumentMetadataFilter

DocumentMetadataFilter(;
    document_id=nothing,
    source=nothing,
    source_id=nothing,
    author=nothing,
    start_date=nothing,
    end_date=nothing,
)

- document_id::String
- source::Source
- source_id::String
- author::String
- start_date::String
- end_date::String

GptSearchPlugin.AppServer.GptPluginServer.HTTPValidationError — Type

HTTPValidationError

HTTPValidationError(;
    detail=nothing,
)

- detail::Vector{ValidationError}

GptSearchPlugin.AppServer.GptPluginServer.LocationInner — Type

Location_inner

LocationInner(; value=nothing)

GptSearchPlugin.AppServer.GptPluginServer.Query — Type

Query

Query(;
    query=nothing,
    filter=nothing,
    top_k=3,
)

- query::String
- filter::DocumentMetadataFilter
- top_k::Int64

GptSearchPlugin.AppServer.GptPluginServer.QueryRequest — Type

QueryRequest

QueryRequest(;
    queries=nothing,
)

- queries::Vector{Query}

GptSearchPlugin.AppServer.GptPluginServer.QueryResponse — Type

QueryResponse

QueryResponse(;
    results=nothing,
)

- results::Vector{QueryResult}

GptSearchPlugin.AppServer.GptPluginServer.QueryResult — Type

QueryResult

QueryResult(;
    query=nothing,
    results=nothing,
)

- query::String
- results::Vector{DocumentChunkWithScore}

GptSearchPlugin.AppServer.GptPluginServer.QueryWithEmbedding — Type

QueryWithEmbedding

QueryWithEmbedding(;
    query=nothing,
    filter=nothing,
    top_k=3,
    embedding=nothing,
)

- query::String
- filter::DocumentMetadataFilter
- top_k::Int64
- embedding::Vector{Float64}

GptSearchPlugin.AppServer.GptPluginServer.UpsertRequest — Type

UpsertRequest

UpsertRequest(;
    documents=nothing,
)

- documents::Vector{Document}

GptSearchPlugin.AppServer.GptPluginServer.UpsertResponse — Type

UpsertResponse

UpsertResponse(;
    ids=nothing,
)

- ids::Vector{String}

GptSearchPlugin.AppServer.GptPluginServer.ValidationError — Type

ValidationError

ValidationError(;
    loc=nothing,
    msg=nothing,
    type=nothing,
)

- loc::Vector{LocationInner}
- msg::String
- type::String

GptSearchPlugin.AppServer.GptPluginServer.register — Method

Register handlers for all APIs in this module in the supplied Router instance.

Paramerets:

router: Router to register handlers in

impl: module that implements the server methods

Optional parameters:

path_prefix: prefix to be applied to all paths

optional_middlewares: Register one or more optional middlewares to be applied to all requests.

Optional middlewares can be one or more of: - init: called before the request is processed - pre_validation: called after the request is parsed but before validation - pre_invoke: called after validation but before the handler is invoked - post_invoke: called after the handler is invoked but before the response is sent

The order in which middlewares are invoked are: init |> read |> pre_validation |> validate |> pre_invoke |> invoke |> post_invoke